17_djexpm/historical_facts/requirements.txt
Django
Pillow
djangorestframework
django-filter
django-cors-headers

